+++ src/server/v2/index.ts
@@ -2,3 +2,4 @@ export * from "./BaseMcpServer.js";
 export * from "./StdioServer.js";
 export * from "./StreamableHttpServer.js";
 export * from "./SseServer.js";
+export * from "./IHttpApplication.js";